refactor: extract organize plan service
